Removed download-dependencies-step to fix docker build
Had to remove the download-dependencies-step because of the following:
While this works OK for 1-module builds, this does not really work in the case of dependencies between modules since no jar will be created.
Maven will not find the module that is depend upon in the local repository (no sources => no archive), and will try to download the dependency.
But since the build is not yet uploaded to nexus in case of releases, it won't be able to find it there, causing an error.