Added report project/profile to fix coverage reporting
SonarQube seems to require an additional project specifically for reporting.
This project must have the dependencies to the projects to report.
This change adds this project, but only under the profile sonar,
which is only used when running SonarQube.

Changed name in parent pom to show nicely in SonarQube.

Added badges to README