Changes the way workers monitor the available workers
Instead of using the RabbitMQ admin api this creates a new fanout exchange.
To this exchange for each type of worker the size and utilisation is published.
Workers can then use these metrics to determine the how many tasks can be done in parallel.

Part of this is the taskmanager listening to the rabbitmq events to get quick feedback on when workers are added and removed.
For this the plugin rabbitmq_event_exchange must be enabled with RabbitMQ.
Without it would still work, it will just update the worker state on the configured refresh time period.

Also some code improvements are included in this change.