AER-77 Fix handling shutdown of RabbitMQ connection (#27)
When the RabbitMQ connection shuts down the tasks tracked in the Task Manager will be zombies as they will not be reported back by the workers as to be finished because due to the connection loss they will be put back on the queue.
Therefor when the connection loss happens the Task Manager should clean up internally all tasks it tracks and restart connections and wait for the tasks to re-appear.

Changed that TaskDispatcher forwardTask locks after setting task on the scheduler.
Before it claimed a lock, put on the scheduler and returned. Then when the next task was retrieved from the queue and the task was forwarded it locked at that point.
This was changed because when all tasks are killed we now know that no tasks are halfway in the forwardTask method.
Because previous when tasks where killed, the task waiting in the forwardTask method could passed on to the scheduler when kill was in action due to timing issues.