AER-274 Make taskmanager configurable via environment
At this point, the task manager gets its configuration from files. As soon as there is a configuration change in one of these files, the task manager will automatically reload its configuration. Modifying files is just not very useful in Docker / containers, it is better to modify it via environment variables, so that it is also possible to easily adapt these per environment. That is more consistent compared to other configuration options in AERIUS.

The configuration will remain in the files. If an environment variable exists for a particular worker type, the settings will be overwritten by this configuration. The value in the environment variable takes precedence over the value in the file.