Rare nullpointer-due-to-racecondition fix (#38)
In rare cases, a nullpointer could occur when unlocking a client (unLockClient method). This could be due to a race condition. The thread where the task would be picked up and released (the run method) can be faster in doing it's work than the original thread that added the task to the scheduler and which would lock after doing so. Since the semaphore would only be added to the map when locking, that could result in a nullpointer on the run thread. By ensuring the lock is added to the map before adding it to the scheduler, we should prevent this.