Switched to lock+condition in PriorityTaskScheduler
The semaphore approach with a 'hasQueuedThreads' check was a bit clunky, and had a chance on a race condition: if taskmanager was started in debug, with breakpoints on the acquire and the release (the 'hasQueuedThreads' line) bits, and first all threads with the release would be continued, and then the acquire thread, the system would basically stop working. The semaphore would never be released again, unless a worker update would occur. This could have been fixed by using the 'tryAcquire' method with time parameters. However, don't think the semaphore object is a right fit here: it's meant as a replacement for a Thread.sleep, so a Condition that gets signalled makes more sense. It's also tied to a Lock, which means we can do the waiting within the synchronized block (which have now been replaced by the lock mechanism).
The 2 minute wait time is a fallback, in case we missed something: if we forgot to signal somewhere, this class will still periodically check if there's something to do, just to avoid 'deadlocks'.