AER-879 Apply GZIP compression on message payloads
This compresses all payloads we pass using RabbitMQ with Gzip. Local experimentation shows that this reduces the payloads we generate using Java serialization to about 25% of the original size, with large payloads (large number of `EngineSource`s) reducing consistently to about 7% of the original size.

This greatly reduces the IO and memory/diskspace required by the taskmanager and message queue, while only very slightly increasing the load on publishers or consumers of tasks.