Avoid using executorservice to start reply consumer
As it was, the reply consumer wouldn't be started until 10 seconds have passed (the DEFAULT_RETRY_SECONDS as used here is actually a DELAY_SECONDS...).
Because we don't want to start the actual queue consumer before the reply consumer has been started, it doesn't make sense to use an executorService: this shouldn't be started asynchronously.
This should fix some issues where a task was handled before the reply queue was actually created, which in turn would cause the taskmanager to never get a signal that the task was finished.