AER-1413 Improved stability to handle connection errors
Both starting and the shutdownhandler can get a ShutdownSignalException.
This could result in both the start trying to restart as well as via the TaskConsumer shutdownhandler to get a seconds call to start.
This changes is intended to not get a second start call when a ShutdownSignalException is thrown during start of the consumer.