AER-2293 - Attempt to reduce error log when restarting taskmanager (#65)
On production system when Taskmanager/RabbitMQ is restarted it can happen 2 taskmanagers are run temporary.
Because some queues have unique access requirements the newly started taskmanager could complain it can't access the queues.
The taskmanager should recover from this, but in the mean time will log numerous errors.
This is an attempt to reduce that logging. It was not possible to reproduce the exact logging from production, but this should at least reduce some logging.