Removed delay from handleShutdownMessage
This caused the taskmanager to stay in an 'old' state as it would block all the other shutdown listeners till all the handlers were done being notified. As there are quite some handlers this would cause the disrupted connection thread to linger way longer than needed.
Also avoiding a sleep when we're shut down anyway. Not entirely sure if this caused issues, but there's not much point in doing so.