Sturdier workermonitor
As it was, an exception could occur in the start method and that would cause the monitor to stop working when RabbitMQ was restarted. As this monitor is used to track other worker utilitzation to dynamically scale, that would grind to a halt as it would never be updated again.
By making it sturdier (catching some exceptions and retrying) the monitor should be better at self-healing when connection with rabbitmq is restored.
Changes are similar to how RabbitMQMessageHandler is implemented.