Taskmanager stability issues (#67)
* Ensure getting connection occurs within try-with-resources
The `factory.getConnection()` call could also result in a `ShutdownSignalException`, doing it this way ensures the loop is not broken when that happens.
* Removed delay from handleShutdownMessage
This caused the taskmanager to stay in an 'old' state as it would block all the other shutdown listeners till all the handlers were done being notified. As there are quite some handlers this would cause the disrupted connection thread to linger way longer than needed.
Also avoiding a sleep when we're shut down anyway. Not entirely sure if this caused issues, but there's not much point in doing so.
* Sturdier workermonitor
As it was, an exception could occur in the start method and that would cause the monitor to stop working when RabbitMQ was restarted. As this monitor is used to track other worker utilitzation to dynamically scale, that would grind to a halt as it would never be updated again.
By making it sturdier (catching some exceptions and retrying) the monitor should be better at self-healing when connection with rabbitmq is restored.
Changes are similar to how RabbitMQMessageHandler is implemented.