Doing something with boolean returned by await method (#68)
Sonar indicated this as a bug. It's not really the way we're using it: we don't really care if there was a signal or if the wait time is over, we act the same anyway. But for debugging purposes it might be nice to log this, so added this information as a trace.