Add mockito as test dependency (#84)
Root pom now specifies mockito as a javaagent. In most cases that's fine, but some modules do not yet have mockito as a test dependency, and would fail because the necessary property with location of the javaagent jar would not be present.
As it's a common test dependency, think it's fine to add it even if mockito isn't (yet) used in a module.