Move scheduler implementation to own package
This to keep better oversight of the code as more features will be added to the scheduler.
Also moved some code out of PriorityScheduler in it's own class.
This class is in preparation of dynamic queue creation.

Also set library version to 2.0.0-SNAPSHOT to indicate the changes (and more that are in the pipeline) are major.