Move scheduler implementation to own package (#95)
This to keep better oversight of the code as more features will be added to the scheduler. Also moved some code out of PriorityScheduler in it's own class. This class is in preparation of dynamic queue creation. Made TaskConsumer an interface, to avoid having cycles in the dependencies. Moved FifoScheduler to test class and renamed to better make clear its purpose is just for testing.

Also set library version to 2.0.0-SNAPSHOT to indicate the changes (and more that are in the pipeline) are major.