Added eager fetching option to queues
Eager fetching will get all messages from the queue to be send to a worker instead of 1 message at a time.
It still acks the message when actually send to the worker.
With eager fetching it will be possible to apply priority to messages on the same queue.
This way we can get a more fair scheduler. Since it will give higher priority to messages from the same job that have less tasks already running on a worker.
With fairer priority the time a calculation takes should become more consist on average.
Getting all messages from the queue can take some more memory, but with plans for a new way to put messages on the queue
the actual messages put on the queues should become limited.

This new feature is backward compatible. As long as eagerFetch is not enabled it will work as before.

Some other refactorings:
- Removed the locking of the queues from TaskDispatcher. For eager fetching no locking should be done.
  But it turned out RabbitMQ already does the locking via QoS value. Therefore the locking was actually redundant.
- Removed MessageMetaData. It's only used for RabbitMQ deliveryTag that can be put int RabbitMQMessage, making the overall code simpler.
- Refactored some mocking classes by using Mockito instead of mocked class.