Changed Historgram metrics to gauge metrics (#103)
Historgram metrics show up with past taking into account, not given an accurate value of the state at the moment these values are not helpful currently.
Therefore they are changed to gauge metrics. Also because the values itself are already aggregated in the taskmanager it does make more sense to have gauge values.