AER-4020 fix for misscounting the load metric
When the taskmanager is restarted and there were still tasks on the queue the PerformanceMetricsReport will get these tasks and compute them even if the taskmanager doesn't know the metric.
Because after restart it has no information on the tasks that were still on the worker queue. Therefore it should keep track on what it has put on the queue and only call register on tasks it knows and ignore tasks that it doesn't know about.
It can compute the duration because that is based on information in the task which it receives from the queue and not from information in the taskmanager itself.

Added safeguard to LoadMetric that it should never go negative. If the calculation would go to negative it means there were more tasks completed than have been accounted for.
By making sure it won't go negative it will correct itself for when task would be reported that were not known (for example after restart).
Technically with the other change in the commit it should never get to negative as tasks unknown would not trigger register call.