AER-4020 fix for miss-counting the load metric (#109)
When the taskmanager is restarted and there were still tasks on the queue the PerformanceMetricsReport will get these tasks and compute them even if the taskmanager doesn't know the metric. Because after restart it has no information on the tasks that were still on the worker queue. Therefore it should keep track on what it has put on the queue and only call register on tasks it knows and ignore tasks that it doesn't know about. It can compute the duration because that is based on information in the task which it receives from the queue and not from information in the taskmanager itself.